# swagger/swagger.yml
openapi: 3.0.1
info:
  title: GeoPlaces API
  description: GeoPlaces API for Node Server
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Authentication server (Local)
  - url: http://localhost:3000/api/v1
    description: API server v1 (Local)
...
paths:
...
  /auth/refresh-token:
    post:
      tags:
        - Auth
      summary: Refresh Token
      operationId: RefreshToken
      requestBody:
        content:
          application/json:
            schema:
              properties:
                refreshToken:
                  type: string
                  description: Refresh token
      responses:
        "200":
          description: Successfull Response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/IAPIResponse"
                  - properties:
                      body:
                        type: object
                        properties:
                          token:
                            type: string
  /users/{id}:
    get:
      tags:
        - User
      summary: Get User by Id
      description: Returns a single user
      operationId: getUserById
      parameters:
        - name: id
          in: path
          description: UID of user to return
          required: true
          schema:
            type: string
            format: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/IAPIResponse"
                  - properties:
                      body:
                        type: object
                        properties:
                          user:
                            $ref: "#/components/schemas/User"
                          token:
                            type: string
                          refreshToken:
                            type: string
        "401":
          $ref: "#/components/responses/Unauthorized" # <-----
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - JWT: []
...
components:
  securitySchemes:
    JWT:
      type: apiKey
      name: Authorization
      in: header
...
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/IAPIErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/IAPIErrorResponse"
  schemas:
    IAPIErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          readOnly: true
          default: false
        message:
          type: string
          readOnly: true
        error:
          type: object
          readOnly: true
      required:
        - code
        - message
    IAPIResponse:
      type: object
      properties:
        success:
          type: boolean
          readOnly: true
        message:
          type: string
          readOnly: true
        body:
          oneOf:
            - type: object
          readOnly: true
    User:
      allOf:
        - properties:
            uid:
              type: string
              format: uuid
            full_name:
              type: string
            firstName:
              type: string
            lastName:
              type: string
            email:
              type: string
              nullable: true
              format: email
            status:
              type: integer
              format: int32
            profileCompletion:
              type: integer
              format: int32
            userStatus:
              type: integer
              format: int32
              description: User Status
...